Rename styles.ts for Nav components
Rename styles.ts to nav.styles.ts
- update import paths as needed
- add **.styles.ts to the ignore list in jest.config.js
