+++ client/src/containers/Nav/FiltersCard.tsx
@@ -10,7 +10,7 @@ import ListItemText from '@material-ui/core/ListItemText';
 import Box from '@material-ui/core/Box';
 
 
-import { filterCardStyles as useStyles } from './styles';
+import { filterCardStyles as useStyles } from './nav.styles';
 
 const FiltersCard: FunctionComponent<any> = (props) => {
   const classes = useStyles();
