+++ jest.config.js
@@ -28,6 +28,7 @@ module.exports = {
   // An array of glob patterns indicating a set of files for which coverage information should be collected
   collectCoverageFrom: [
     "**/*.ts",
+    "!**/*.styles.ts",
   ],
 
   // The directory where Jest should output its coverage files
@@ -156,7 +157,7 @@ module.exports = {
 
   // An array of regexp pattern strings that are matched against all test paths, matched tests are skipped
   testPathIgnorePatterns: [
-    "\\\\node_modules\\\\"
+    "\\\\node_modules\\\\",
   ],
 
   // The regexp pattern or array of patterns that Jest uses to detect test files
